back/requirements.txt
# Django Core
Django==5.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.3
python-decouple==3.8

# Database drivers - instalar según necesidad
# PostgreSQL (predeterminado)
psycopg2-binary==2.9.9
drf-spectacular==0.27.2

# MySQL/MariaDB (opcional)
# Descomenta para usar MySQL/MariaDB
# pymysql==1.1.0
# mysqlclient==2.2.0

# Oracle (opcional)
# Descomenta para usar Oracle
# cx-Oracle==8.3.0

# Herramientas adicionales
Pillow==10.0.1  # Para campos ImageField
python-dateutil==2.8.2  # Para manejo de fechas
